Add base renderer
